More edge cases

+++ spec/spec_helper.rb
@@ -15,7 +15,7 @@ require_relative "../lib/whedon"
 # running tests (spec/whedon_spec.rb#L11). Also, we're hard-coding the path
 # to tmp/#{review_issue_id} so we need a fixture there.
 Dir.mkdir('tmp') unless Dir.exist?('tmp')
-FileUtils.rm_r('tmp/17')
+FileUtils.rm_r('tmp/17') if Dir.exist?('tmp/17')
 FileUtils.copy_entry('fixtures/paper', 'tmp/17')
 
 # Any git repo will do (using the one from Whedon)
